Update IntakeSystem.java
Removed the variable atributed as "noteInIntake", due to redundancy in the intake system. Now, the code compiles one hundred thousandth of a second faster, and the overall performance of the robot is improved by one hundred thousandth of a second. This should improve our ability to compete in the competition.